Rename 'start-packs' files to 'resources'
