+++ app/client/src/app/views/resources/project-ideas/projectIdeas.scss
