+++ app/client/src/app/views/resources/resources.scss
@@ -3,12 +3,6 @@
 .starter-packs {
   min-height: calc(100vh - 114px);
   width: 100vw;
-  background: linear-gradient(
-    180deg,
-    #8c765f 0.8%,
-    #8a5655 48.2%,
-    #555776 100%
-  );
   display: flex;
   align-items: center;
   justify-content: center;
